Refactor CurrentStorage to delegate vector logic to VectorIndexManager
- Extracted `VectorIndexManager` to `src/storage/current/vector.rs`.
- Moved `vector_indexes`, `temporal_vector_indexes`, and `filter_stats` to `VectorIndexManager`.
- Updated `CurrentStorage` to own `VectorIndexManager` and delegate calls.
- Fixed unrelated clippy warnings in `src/experimental/semantic_pathfinding.rs`.
- Fixed CI flakiness in `tests/issue_209_version_lookup_performance.rs` by disabling persistence.
